requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cftime==1.6.4.post1
chardet==3.0.4
cloudinary==1.44.1
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.9
dj-database-url==3.0.1
Django==5.1.2
django-cors-headers==4.6.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
filelock==3.16.1
fonttools==4.54.1
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kiwisolver==1.4.7
matplotlib==3.9.2
MouseInfo==0.1.3
mysqlclient==2.2.6
netCDF4==1.7.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2024.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.5.0
virtualenv==20.28.0
xarray==2024.10.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url>=2.1.0